CA-136792: Read VBD IO metrics from '/dev/shm'.
With blktap3, blkback is not on the datapath except for rare occasions.
The statistics originally produced by blkback are now done from tapdisk.
This patch first attempts to read the statistics from under '/dev/shm'
to produce RRDs and defaults to the original sysfs statistics tree when
there are no valid files under 'dev/shm'.

Signed-off-by: Siddharth Vinothkumar <siddharth.vinothkumar@citrix.com>